:+1: adjust GameInfo
