+++ skeleton-app/src/routes/(auth)/simulator/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import DuelField from "$lib/components/organisms/DuelField.svelte";
-  import PlayerInfo from "$lib/components/organisms/PlayerInfo.svelte";
+  import GameInfo from "$lib/components/organisms/GameInfo.svelte";
   import Hands from "$lib/components/organisms/Hands.svelte";
 
   let gameState = {
@@ -18,20 +18,18 @@
 
 <div class="container mx-auto p-4">
   <main class="max-w-6xl mx-auto space-y-6">
-    <DuelField
-      deckCards={gameState.deckCards}
-      extraDeckCards={gameState.extraDeckCards}
-      graveyardCards={gameState.graveyardCards}
-    />
-
-    <Hands handCards={gameState.handCards} />
-
-    <PlayerInfo
+    <GameInfo
       playerLifePoints={gameState.playerLifePoints}
       opponentLifePoints={gameState.opponentLifePoints}
       currentTurn={gameState.currentTurn}
       currentPhase={gameState.currentPhase}
       gameStatus={gameState.gameStatus}
     />
+    <DuelField
+      deckCards={gameState.deckCards}
+      extraDeckCards={gameState.extraDeckCards}
+      graveyardCards={gameState.graveyardCards}
+    />
+    <Hands handCards={gameState.handCards} />
   </main>
 </div>
